Add "user specific" debugging APIs.
The APIs from DEBUG_WOLFSSL work great, but can take up a lot of space when enabled.  On platforms with limited resources, it may not be possible to enable DEBUG_WOLFSSL and have the image still fit.

USER_DEBUG_WOLFSSL is intended to only be used when a specific issue is being debugged.  It can replace the few important WOLFSSL_MSG or WOLFSSL_ERROR calls in the code path desired or can be used to add a few more debug statements where the developer needs it without getting the whole DEBUG_WOLFSSL output.