Create a new function, ParseCertLockOptional, that's equivalent to ParseCert, but knows if the CA table lock is held.
This resolves an issue where ParseCert was called by the function
wolfSSL_CertManagerGetCerts after that function acquired the lock in question.
Ultimately, ParseCert calls GetCA, which attempts to acquire the same lock.
Deadlock ensues. By passing a parameter down that indicates the lock is already
held by the calling function, GetCA knows whether or not it needs to acquire the
lock itself.