Fix alignment in ‘wolfSSL_X509_VERIFY_PARAM_set1_host’
Reported by GCC:

src/ssl.c: In function ‘wolfSSL_X509_VERIFY_PARAM_set1_host’:
src/ssl.c:27315:5: error: this ‘if’ clause does not guard... [-Werror=misleading-indentation]
27315 |     if (nameSz > 0)
      |     ^~
src/ssl.c:27318:9: note: ...this statement, but the latter is misleadingly indented as if it were guarded by the ‘if’
27318 |         pParam->hostName[nameSz] = '\0';
      |         ^~~~~~

The assignment of the null terminator is aligned with the preceding if
block, butt here are no brackets. Thus GCC complains about the ambiguity
(or potential error).

Fix indentation by moving the instruction to the proper column.

Signed-off-by: Antonio Quartulli <a@unstable.cc>