Fix configure.ac to generate empty files
The fixes 'make dist' which is currently broken. To confirm error that this fixes:
autoreconf -if
./configure
make dist

FWIW normally repositories would use *.in to generate the above named files in an empty manner instead of using a blank ":" operator. I don't know how you use these files so I am unsure if you want that since the *.in would need to have rules around it being shipped as well.
All rules right now that would force autoreconf will also fail.
The usage of a "autogen.sh" was discontinued by autoconf a number of years ago in favor of just using autoreconf.